hiberman: Store metadata header size in bytes
The upcoming kernel hardening solution uses AES-GCM to have the kernel
encrypt the image, and interleaves chunks of data with 16-byte
authentication tags. This means the header and data size may not be an
even number of pages.

In order to accurately store the header region size in the future,
change the metadata pagemap_pages into image_meta_size. This represents
the number of bytes in the header page and pagemap.

While in there, change the total image_size (which was already in
bytes) from a u64 to an i64, since that's the type returned by the
kernel (via loff_t).

BUG=b:206143128
TEST=cargo build

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3603151
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>

+++ hiberman/src/suspend.rs
@@ -250,7 +250,7 @@ impl SuspendConductor {
 
         assert!(self.metadata.data_tag != [0u8; META_TAG_SIZE]);
 
-        self.metadata.image_size = image_size as u64;
+        self.metadata.image_size = image_size;
         self.metadata.flags |= META_FLAG_VALID;
         Ok(())
     }
